requirements.txt
flask
python-dotenv
google-generativeai
faiss-cpu
sentence-transformers
transformers
huggingface-hub
pandas
langchain
